+++ sdk/src/azure/core/az_http_policy.c
@@ -4,13 +4,13 @@
 #include "az_http_private.h"
 #include <azure/core/az_credentials.h>
 #include <azure/core/az_http.h>
+#include <azure/core/az_precondition.h>
 #include <azure/core/az_span.h>
+#include <azure/core/az_version.h>
 #include <azure/core/internal/az_http_internal.h>
 #include <azure/core/internal/az_result_internal.h>
 
 #include <azure/core/_az_cfg.h>
-
-static const az_span AZ_HTTP_HEADER_USER_AGENT = AZ_SPAN_LITERAL_FROM_STR("User-Agent");
 
 AZ_NODISCARD az_result az_http_pipeline_policy_apiversion(
     _az_http_policy* ref_policies,
@@ -42,20 +42,43 @@ AZ_NODISCARD az_result az_http_pipeline_policy_apiversion(
   return _az_http_pipeline_nextpolicy(ref_policies, ref_request, ref_response);
 }
 
+#define _az_TELEMETRY_ID_PREFIX "azsdk-c-"
+#define _az_TELEMETRY_ID_PREFIX_LENGTH (sizeof(_az_TELEMETRY_ID_PREFIX) - 1)
+
 AZ_NODISCARD az_result az_http_pipeline_policy_telemetry(
     _az_http_policy* ref_policies,
     void* ref_options,
     az_http_request* ref_request,
     az_http_response* ref_response)
 {
+  _az_PRECONDITION_NOT_NULL(ref_options);
 
-  _az_http_policy_telemetry_options* options = (_az_http_policy_telemetry_options*)(ref_options);
+  // Format spec: https://azure.github.io/azure-sdk/general_azurecore.html#telemetry-policy
+  uint8_t telemetry_id_buffer[200] = _az_TELEMETRY_ID_PREFIX;
+  az_span telemetry_id = AZ_SPAN_FROM_BUFFER(telemetry_id_buffer);
+  {
+    az_span remainder = az_span_slice_to_end(telemetry_id, _az_TELEMETRY_ID_PREFIX_LENGTH);
+
+    _az_http_policy_telemetry_options* options = (_az_http_policy_telemetry_options*)(ref_options);
+    az_span const component_name = options->component_name;
+    _az_PRECONDITION_VALID_SPAN(component_name, 1, false);
+    remainder = az_span_copy(remainder, component_name);
+
+    remainder = az_span_copy_u8(remainder, '/');
+    remainder = az_span_copy(remainder, AZ_SPAN_FROM_STR(AZ_SDK_VERSION_STRING));
+
+    telemetry_id = az_span_slice(
+        telemetry_id, 0, (int32_t)(az_span_ptr(remainder) - az_span_ptr(telemetry_id)));
+  }
 
   _az_RETURN_IF_FAILED(
-      az_http_request_append_header(ref_request, AZ_HTTP_HEADER_USER_AGENT, options->os));
+      az_http_request_append_header(ref_request, AZ_SPAN_FROM_STR("User-Agent"), telemetry_id));
 
   return _az_http_pipeline_nextpolicy(ref_policies, ref_request, ref_response);
 }
+
+#undef _az_TELEMETRY_ID_PREFIX
+#undef _az_TELEMETRY_ID_PREFIX_LENGTH
 
 AZ_NODISCARD az_result az_http_pipeline_policy_credential(
     _az_http_policy* ref_policies,