+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 - Added `az_http_response_get_status_code()` convenience function to get HTTP status code from requests.
+- Improved HTTP request telemetry.
 
 ### Breaking Changes